rename SPIRIT_PO_NOEXCEPT -> SPIRIT_PO_NO_EXCEPTIONS
The old name is actually confusing. It doesn't make things noexcept.
It makes them not throw exceptions intentionally.

They are not actually noexcept -- they make dynamic allocations, for
instance.

The flag is only useful for compiling with -fno-exceptions. It's
better to mirror the boost nomenclature for this preprocessor symbol.

I have added a line in catalog.hpp so that if we see `SPIRIT_PO_NOEXCEPT`
we define `SPIRIT_PO_NO_EXCEPTIONS` but I'm only going to document the new
name.