minor source improvements
defensive programming against possibility of bad plural forms expressions

we make it so that when a parse error occurs with the plural form expression,
the "default" function that we provide is constant zero function, rather than
`n`. The reason is that n is likely to cause out of bounds accesses, while 0
isn't.

We also test more carefully for those out of bounds accesses, to avoid UB
if the plural forms expression produces bad output.

We also make sure to always initialize the "singular_index_" variable, since
that could cause UB also if the user avoided warnings / errors in a catalog
during it's construction and then proceeded to use it anyways.